COMMIT EMIL ELDOOEN 22.02.2021
Added following:
Added dealer class to deal cards to players
Has methods:
deal(), returns a card from the top of the stack
getDeck() allows us to fetch the card deck
shuffleDeck() allowing us to easily shuffle the dealer's deck.

Made tests for the dealer class. These tests checks if the dealer's card is removed from the stack once the deal function has been called and that the sizes of the deck is the same once sent into the dealer class.